invoices: move InvoiceDB tests unrelated to kvdb to invoices package
