+++ invoices/setup_test.go
@@ -1,0 +1,11 @@
+package invoices
+
+import (
+	"testing"
+
+	"github.com/lightningnetwork/lnd/kvdb"
+)
+
+func TestMain(m *testing.M) {
+	kvdb.RunTests(m)
+}
